Add authn/authz to permission.delete()

+++ server/api/permission/delete/service.js
@@ -1,26 +1,16 @@
-import { PermissionNotFoundError, ImmutablePermissionError } from '../../../constants/errors';
+import { ImmutablePermissionError } from '../../../constants/errors';
 
-async function deletePermission(db, { id }) {
+async function deletePermission(db, permission) {
   const PermissionModel = db.model('Permission');
-
-  // retrieve permission from db
-  const permission = await PermissionModel.findOne({
-    _id: id,
-  });
-
-  // make sure permission exists
-  if (!permission) {
-    throw new PermissionNotFoundError(`Permission ${id} cannot be found`);
-  }
 
   // make sure permission is not system-defined
   if (permission.isSystemPermission) {
     throw new ImmutablePermissionError(
-      `Permission ${id} is a system permission and thus cannot be deleted`
+      `Permission ${permission.id} is a system permission and thus cannot be deleted`
     );
   }
 
-  const result = await PermissionModel.deleteOne({ _id: id });
+  const result = await PermissionModel.deleteOne({ _id: permission.id });
   return !!result.ok;
 }
